+++ src/main/java/pe/edu/utp/BibMpch/repository/GenderRepository.java
@@ -3,5 +3,5 @@ package pe.edu.utp.BibMpch.repository;
 import org.springframework.data.repository.CrudRepository;
 import pe.edu.utp.BibMpch.model.Gender;
 
-public interface GenderRepository extends CrudRepository<Gender, Long> {
+public interface GenderRepository extends CrudRepository<Gender, Short> {
 }
